Change all instances of sps to plantasy
