+++ src/main/java/com/google/plantasy/servlets/UploadHandlerServlet.java
@@ -12,43 +12,48 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
  
-package com.google.sps.servlets;
+package com.google.plantasy.servlets;
  
+import com.google.appengine.api.blobstore.BlobKey;
+import com.google.appengine.api.blobstore.BlobstoreService;
+import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import com.google.sps.utils.GameUtils;
-import com.google.sps.utils.UserUtils; 
+import com.google.plantasy.utils.UserUtils;
 import java.io.*;
 import javax.servlet.*;
-import javax.servlet.http.*;
-import java.io.PrintWriter;
+import javax.servlet.http.*; 
  
-/** When a user joins a game, check if the given game id exists
-    if so, get the game entity and add the user,
-    then add the game to the user entity.
-    If it doesn't exist, redirect back to the join game page.
-*/
-@WebServlet("/join-game")
-public class joinGameServlet extends HttpServlet {
+/** Blobstore upload handler: gets blobkey (Blobstore rewrites the request to contain a blobkey) 
+    for an uploaded image and stores it in datastore 
+    (see https://cloud.google.com/appengine/docs/standard/java/blobstore#3_implement_upload_handler
+    for documentation of an upload handler)
+ */
+@WebServlet("/image-upload-handler-blobstore")
+public class UploadHandlerServlet extends HttpServlet {
  
+  BlobstoreService blobstoreService;
   DatastoreService datastore;
  
-  public joinGameServlet(){
+  public UploadHandlerServlet(){
+    blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
     datastore = DatastoreServiceFactory.getDatastoreService();
   }
  
+  // TODO: Limit this so it can only be used as a blobstore upload handler; cut off end-user access
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
     
-    String gameId = request.getParameter("game-id");
- 
-    // get user entity
+    // get the user entity 
     Cookie cookies[] = request.getCookies();
     if(cookies == null){
         response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
@@ -59,28 +64,14 @@ public class joinGameServlet extends HttpServlet {
         response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
         return;
     }
-
-    // prevent users from joining more than one game
-    if(userEntity.getProperty("gameId") != null){
-      response.sendRedirect("/gameBoard.html");
-      return;
-    }
-
-    Entity gameEntity = UserUtils.getEntityFromDatastore("Game", "gameId", gameId, datastore);
-    if(gameEntity = null){
-      PrintWriter out = response.getWriter();
-      out.println("<p>We couldn't find a game by that code, <a href = \"join.html\"><h3>press here</h3></a> and enter a different game code.</p>");
-      return;
-    }
-    boolean setGame = GameUtils.setGame(userEntity, datastore, newGame);
-
-    if(!setGame){
-      // connecting the game to the user failed because the user was not logged in, send back to login
-      response.sendRedirect("/index.html");
-      return;
-    }else{
-      response.sendRedirect("/gameBoard.html");
-      return;
-    }
+    
+    // getting the blobkey + making it part of the request
+    ImmutableMap<String, List<BlobKey>> blobs = ImmutableMap.copyOf(blobstoreService.getUploads(request));
+    ImmutableList<BlobKey> blobKeys = ImmutableList.copyOf(blobs.get("image"));
+    String blobKey = blobKeys.get(0).getKeyString();
+ 
+    UserUtils.addBlobKey(blobKey, userEntity, datastore);
+    
+    response.sendRedirect("/imageUpload.html"); 
   }
 }
