+++ src/main/java/com/google/plantasy/utils/QuizTimingPropertiesUtils.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
  
-package com.google.sps;
+package com.google.plantasy.utils;
  
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.DatastoreService;
