+++ src/main/java/com/google/plantasy/utils/UserUtils.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.utils;
+package com.google.plantasy.utils;
 
 import java.io.*;
 import javax.servlet.*;
