+++ src/test/java/com/google/plantasy/TimingPropertiesTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
  
-package com.google.sps;
+package com.google.plantasy;
  
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
@@ -25,7 +25,7 @@ import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import com.google.sps.QuizTimingPropertiesUtils;
+import com.google.plantasy.utils.QuizTimingPropertiesUtils;
  
 @RunWith(JUnit4.class)
 public final class TimingPropertiesTest {
